% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_filters.R
\name{apply_filters}
\alias{apply_filters}
\title{Apply Groups of Filter Criteria from Configuration}
\usage{
apply_filters(data, groups, length_column = NULL, id_columns = names(data)[1L])
}
\arguments{
\item{data}{data frame}

\item{groups}{names of filter criteria groups defined in list returned by
\code{kwb.prep:::read_filter_criteria}}

\item{length_column}{name of column in \code{data} containing lengths (to be
summed up for the overview that is returned)}

\item{id_columns}{names of column(s) in \code{data} that uniquely identify
the records. This column / these columns are returned in order to report
about the records that have been removed}
}
\value{
data, filtered according to the specified criteria. The returned data
frame has an attribute \code{filter_info} being a list with as many
elements as there are \code{groups}. The elements are named according to
the values given in \code{groups}. Each list element is a list with one
element \code{overview} (being a data frame with one row per filter
criterion) and further elements \code{removed_<i>} being data frames with
only \code{id_columns} that represent the records that have been removed in
the according filter step \code{i}.
}
\description{
Apply Groups of Filter Criteria from Configuration
}
\examples{

# Define filter criteria
criteria <- list(
  sepal = c(
    "sepal short" = "Sepal.Length < 5",
    "sepal narrow" = "Sepal.Width < 3"
  ),
  petal = c(
    "petal short" = "Petal.Length < 5",
    "petal narrow" = "Petal.Width < 3"
  )
)

# Write criteria to temporary yaml file
tdir <- tempdir()
yaml::write_yaml(criteria, file.path(tdir, "filter_criteria.yml"))

# Set path to temporary "config" folder so that kwb.prep knows about it
set_user_config_dir(tdir)

# Apply filter groups "sepal" and "petal" to the iris dataset
result <- apply_filters(iris, c("sepal", "petal"))

# Have a look at the result
str(result)

}
