% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyFilter.R
\name{applyFilter}
\alias{applyFilter}
\title{Apply Filter Criteria from List}
\usage{
applyFilter(data, criteria_list, element, length_column = NULL)
}
\arguments{
\item{data}{data frame}

\item{criteria_list}{list of (named) vectors of character representing
filter criteria}

\item{element}{name of list element to be selected fom \code{criteria_list}}

\item{length_column}{passed to \code{applyFilterCriteria}}
}
\description{
Apply Filter Criteria from List
}
\examples{
criteria_list <- list(
  apple = c("is red or green" = "colour \%in\% c('red', 'green')"),
  banana = c("is not straight" = "! straight")
)

fruit_properties <- data.frame(
  colour = c("green", "red", "yellow"),
  straight = c(TRUE, TRUE, FALSE)
  )
  
applyFilter(fruit_properties, criteria_list, "apple")
applyFilter(fruit_properties, criteria_list, "banana")

}
