% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyFilterCriteria.R
\name{applyFilterCriteria}
\alias{applyFilterCriteria}
\title{Filter Rows from Data Frame Matching Criteria}
\usage{
applyFilterCriteria(x, criteria = NULL, lengthColumn = NULL, ...)
}
\arguments{
\item{x}{data frame}

\item{criteria}{vector of character defining filter criteria to be evaluated
in x}

\item{lengthColumn}{name of the column containing lengths, e.g. "Length_raw"}

\item{\dots}{passed to \code{\link[kwb.utils]{matchesCriteria}}}
}
\description{
Details about criteria applied and number of rows matching each criterion
is returned in the attribute "details.filter"
}
\examples{
# Create a very simple data frame
df <- data.frame(value = 1:10, group = rep(c("a", "b"), 5))

# Show the data frame
df

# Filter for rows meeting two criteria
result <- applyFilterCriteria(df, c(
  "value is below or equal to 5" = "value <= 5", 
  "group is 'a'" = "group == 'a'"
))

# Show the result
result

# Get the evaluation of each criterion in columns
kwb.utils::getAttribute(result, "matches")

}
