% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R, R/utils.R
\name{assign_objects}
\alias{assign_objects}
\title{Provide all Objects of kwb.prep in the Global Environment}
\usage{
assign_objects()

assign_objects()
}
\description{
This function provides all (also non-exported) function definitions of this
package in the Global environment. This is useful for debugging the code
of a function that calls non-exported functions.
}
