% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breaksToIntervalLabels.R
\name{breaksToIntervalLabels}
\alias{breaksToIntervalLabels}
\title{Create Interval Labels from Breaks Vector}
\usage{
breaksToIntervalLabels(breaks, style = 5, ...)
}
\arguments{
\item{breaks}{numeric vector of breaks}

\item{style}{passed to \code{\link{intervalLabel}}}

\item{\dots}{further arguments passed to \code{\link{intervalLabel}}}
}
\description{
Create labels for intervals defined by breaks in different possible styles
}
