% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_table_name.R
\name{check_table_name}
\alias{check_table_name}
\title{Check for Valid Table Name}
\usage{
check_table_name(table_name)
}
\arguments{
\item{table_name}{R object to be checked for usage as a table name}
}
\description{
Check if the argument can be used as a table name
}
\examples{
try(check_table_name(c("more", "than", "one", "string")))
try(check_table_name("one_is_ok"))
}
