% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_zip_extension.R
\name{check_zip_extension}
\alias{check_zip_extension}
\title{Stop if File Name Does not End with Zip Extension}
\usage{
check_zip_extension(file)
}
\arguments{
\item{file}{path to file to check for .zip or .7z file name extension}
}
\value{
The function does not return anything but stops with a clear error
message in case that \code{file} does not end with something that looks
like the file extension of a compressed file.
}
\description{
Stop if File Name Does not End with Zip Extension
}
