% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_berlin.R
\name{checkGrouping}
\alias{checkGrouping}
\title{Compare Two Columns of a Data Frame (Raw Vs Regrouped)}
\usage{
checkGrouping(data, column_raw, column_cat)
}
\arguments{
\item{data}{data frame}

\item{column_raw}{name of column in \code{data} containing original (raw)
values}

\item{column_cat}{name of column in \code{data} containing the result of
regrouping the raw value into categories}
}
\description{
Compare Two Columns of a Data Frame (Raw Vs Regrouped)
}
