% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_berlin.R
\name{checkNumberOfUnique}
\alias{checkNumberOfUnique}
\title{Show Number of Unique Values in Selected Columns}
\usage{
checkNumberOfUnique(data, columns = names(data))
}
\arguments{
\item{data}{data frame}

\item{columns}{names of columns in \code{data} for which to count unique
values}
}
\description{
Show Number of Unique Values in Selected Columns
}
