% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collect}
\alias{collect}
\title{Collect Elements of Sublists}
\usage{
collect(x, element, default = NULL)
}
\arguments{
\item{x}{a list of lists}

\item{element}{name of list element to be collected from each sublist of
\code{x}}

\item{default}{value to be returned for lists that do not have an element
called \code{element}.}
}
\description{
Collect Elements of Sublists
}
\examples{
x <- list(
  list(a = 1, b = 2),
  list(c = 3, a = 4),
  list(d = 5, e = 6)
)

collect(x, "a")
collect(x, "a", default = 99)

}
