% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_tables.R
\name{contained_tables}
\alias{contained_tables}
\title{Names of "Tables" Provided in Zip Archive}
\usage{
contained_tables(zip_file, config)
}
\arguments{
\item{zip_file}{path to zip archive file (.zip or .7z)}

\item{config}{csv configuration mapping table names to file names}
}
\value{
vector of character containing the names of the tables for which
corresponding files have been found in the given zip archive
}
\description{
Names of "Tables" Provided in Zip Archive
}
\examples{
kwb.prep::contained_tables(
  zip_file = kwb.prep:::invalid_zip("five-rows.zip"), 
  config = list(
    pipes = list(file = "Haltungen_SAN_utf8.csv"),
    trees = list(file = "Haltungen_Baeume_utf8.csv"),
    animals = list(file = "no-animals-there.csv")
  )
)
}
