% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_berlin.R
\name{dataFramesToTextMatrix}
\alias{dataFramesToTextMatrix}
\title{Convert List of Data Frames to Character Matrix}
\usage{
dataFramesToTextMatrix(data_frames)
}
\arguments{
\item{data_frames}{list of data frames}
}
\description{
In the character matrix the data frames appear one below the other. Each
data frame has a header and each data frame is separated from the following
data frame by an empty row.
}
\examples{
data_frames <- list(
  data.frame(a = 1:3, b = 2:4),
  data.frame(a = 1:5, b = 2:6, c = 3:7)
)

dataFramesToTextMatrix(data_frames)

}
