% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyFilterCriteria.R
\name{fieldSummary}
\alias{fieldSummary}
\title{Frequency of Value Combinations in Data Frame Columns}
\usage{
fieldSummary(x, groupBy = names(x)[-1L], lengthColumn = "", na = "Unknown")
}
\arguments{
\item{x}{data frame}

\item{groupBy}{vector of character naming the columns (fields) in \code{x}
to be included in the evaluation. Default: names of all columns in \code{x}
except the first one (assuming it could be an ID column).}

\item{lengthColumn}{optional. Name of column in \code{x} to be summed up}

\item{na}{optional. Value to be treated as \code{NA}. Default: "Unknown"}
}
\description{
Frequency of Value Combinations in Data Frame Columns
}
\examples{
n <- 1000L
sample_replace <- function(x, ...) sample(x, size = n, replace = TRUE, ...)
x <- data.frame(
  pipe_id = 1:n,
  material = sample_replace(c("clay", "concrete", "other")),
  age_cat = sample_replace(c("young", "old")),
  length = as.integer(rnorm(n, 50)),
  stringsAsFactors = FALSE
)

fieldSummary(x)
fieldSummary(x, "age_cat")
fieldSummary(x, "material")
fieldSummary(x, "material", lengthColumn = "length")
}
