% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_column_classes2.R
\name{get_column_classes2}
\alias{get_column_classes2}
\title{Get Expected Table Column Types from Config}
\usage{
get_column_classes2(table_name, config, available = NULL)
}
\arguments{
\item{table_name}{name of table for which to lookup column types in config}

\item{config}{configuration object (list)}

\item{available}{optional. Names of available table columns. If given, only
the types of these columns are returned.}
}
\value{
list with one entry per type. Each element contains the names of the
colums with the corresponding type.
}
\description{
What are the expected column types as defined in the configuration?
}
\examples{
config <- list(animals = list(fields = list(
  cats = list(field = "Katzen", type = "integer"),
  dogs = list(field = "Hunde", type = "double"),
  mice = list(field = "Maeuse", type = "integer")
)))

kwb.prep:::get_column_classes2("animals", config)

}
