% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_csv_filenames.R
\name{get_csv_filenames}
\alias{get_csv_filenames}
\title{Get Names of CSV Files Referenced in Config}
\usage{
get_csv_filenames(config, keep_empty = FALSE)
}
\arguments{
\item{config}{configuration object (list) with one entry per "table", each
of which is expected to have an entry "file"}

\item{keep_empty}{logical. Whether or not to keep "file" entries that are
empty ("")}
}
\value{
vector of character with the file names referenced in \code{config}
}
\description{
Get Names of CSV Files Referenced in Config
}
\examples{
config <- list(
  table_a = list(file = "table-a.csv"),
  table_b = list(file = "table-b.csv")
)
get_csv_filenames(config)
}
