% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_lower_extension}
\alias{get_lower_extension}
\title{Lower Case Extension of a File}
\usage{
get_lower_extension(file)
}
\arguments{
\item{file}{file path or file name}
}
\description{
Lower Case Extension of a File
}
\examples{
get_lower_extension("abc.XYZ")
}
