% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_path.R
\name{get_path}
\alias{get_path}
\title{Resolve Path from Path Dictionary in Config Folder}
\usage{
get_path(x = NULL, ...)
}
\arguments{
\item{x}{key to be looked up in the path dictionary}

\item{\dots}{possible key = value assignments to be used to replace
\<placeholders\> in the path that was looked up}
}
\description{
Resolve Path from Path Dictionary in Config Folder
}
