% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_renamings.R
\name{get_renamings}
\alias{get_renamings}
\title{Get List Defining Renamings from Data Frame}
\usage{
get_renamings(from, to = "column", data = NULL, reader = read_csv_file, ...)
}
\arguments{
\item{from}{name of column of \code{data} to take the "from" values
from}

\item{to}{name of column of \code{data} to take the "to" values from}

\item{data}{data frame defining renamings}

\item{reader}{reader function providing \code{data}. Default:
\code{kwb.prep:::read_csv_file}}

\item{\dots}{arguments passed to the \code{reader} function}
}
\value{
list defining renamings as e.g. expected by
\code{\link[kwb.utils]{renameColumns}}
}
\description{
Get list defining renamings in the form of \code{from = to} assignments from
a data frame read by a function that may be specified.
}
