% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_renamings_from_config.R
\name{get_renamings_from_config}
\alias{get_renamings_from_config}
\title{Get List of Renamings from Configuration}
\usage{
get_renamings_from_config(config, table_name, all = TRUE)
}
\arguments{
\item{config}{list with one element per table/csv file}

\item{table_name}{name of list element within \code{config}}

\item{all}{if \code{FALSE} only the fields with property "required = TRUE"
are considered}
}
\value{
list with original names as names and internal names as values. The
list can be used in a call to \code{\link[kwb.utils]{renameColumns}}
}
\description{
Get List of Renamings from Configuration
}
