% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_text.R
\name{get_text}
\alias{get_text}
\title{Get Text Constant}
\usage{
get_text(key = NULL, ..., raw_strings = get_raw_strings())
}
\arguments{
\item{key}{identifier}

\item{\dots}{additional arguments passed to \code{\link{sprintf}}}

\item{raw_strings}{list with raw string definitions as key = value pairs}
}
\value{
if \code{key} is \code{NULL}) a list with all text constants or the
text constant looked up for the given key
}
\description{
Get Text Constant
}
