% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_zipped_paths.R
\name{get_zipped_paths}
\alias{get_zipped_paths}
\title{List Files in Zip Archive}
\usage{
get_zipped_paths(zip_file, include_dirs = FALSE)
}
\arguments{
\item{zip_file}{path to zip archive}

\item{include_dirs}{if \code{TRUE} directory paths are also returned. The
default is \code{FALSE}, i.e. only the paths to files are returned.}
}
\value{
paths to files contained in zip archive
}
\description{
List Files in Zip Archive
}
