% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_berlin.R
\name{getChangesOfDuplicates}
\alias{getChangesOfDuplicates}
\title{Get Changes of Rows That Are Duplicated in Selected Columns}
\usage{
getChangesOfDuplicates(df, columns, add_columns = columns)
}
\arguments{
\item{df}{a data frame}

\item{columns}{names of columns in \code{df} in which to look for duplicate
value combinations}

\item{add_columns}{names of additional columns that shall appear in the
output even if there are no changes in these columns}
}
\value{
list of data frames. The list has as many elements as there are
different value combinations in \code{columns} that appear more than once
in \code{df}. Each element is a data frame with all rows from \code{df}
that have the same value combination in \code{columns}. By default the data
frame contains the columns given in \code{columns} and those columns out of
\code{df} in which there is at least one change over the values in the
different rows.
}
\description{
Get Changes of Rows That Are Duplicated in Selected Columns
}
\examples{
df <- data.frame(
  id = 1:7, 
  name = c("one", "one", "two", "two", "three", "three", "three"), 
  type = c("A", "A", "B", "C", "D", "D", "D"),
  size = c(10, 11, 12, 12, 13, 13, 14),
  height = c(1, 1, 2, 3, 4, 4, 5)
)

df

getChangesOfDuplicates(df, "name")
getChangesOfDuplicates(df, c("name", "type"))

}
