% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_berlin.R
\name{getYearFromColumn}
\alias{getYearFromColumn}
\title{Get Integer Year Number from Column}
\usage{
getYearFromColumn(data, column)
}
\arguments{
\item{data}{data frame}

\item{column}{representing a date or date and time}
}
\value{
vector of integer as long as the number of rows in \code{data}
}
\description{
Get Integer Year Number from Column
}
