% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_zip_extension.R
\name{has_zip_extension}
\alias{has_zip_extension}
\title{Does a File have a Zip Extension (.zip, .7z)?}
\usage{
has_zip_extension(file, expected = c("zip", "7z"))
}
\arguments{
\item{file}{path(s) to file(s) to be checked for zip extension}

\item{expected}{expected file name extensions. Default: \code{c("zip", "7z")}}
}
\value{
vector of logical
}
\description{
Does a File have a Zip Extension (.zip, .7z)?
}
\examples{
all(has_zip_extension(c("a.zip", "b.ZIP", "c.Zip", "d.7z", "e.7Z"))) # TRUE
has_zip_extension("a.txt") # FALSE
}
