% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breaksToIntervalLabels.R
\name{intervalLabel}
\alias{intervalLabel}
\title{Create Interval Label from Upper and Lower Boundary}
\usage{
intervalLabel(a, b, right = TRUE, style = 1, sep = ",", space = " ")
}
\arguments{
\item{a}{upper boundary}

\item{b}{lower boundary}

\item{right}{if TRUE (default) the interval is closed at the upper boundary}

\item{style}{integer number between 1 and 5 indicating one of five possible
styles to name the interval between \code{a} and \code{b}. See examples
below.}

\item{sep}{separator to be used between lower and upper boundary}

\item{space}{space between comparison operators and boundary values.}
}
\description{
Create a label for the interval defined by the upper boundary \code{a} and
the lower boundary \code{b}
}
\examples{
# Labels of different styles for right closed intervals (right = TRUE is the 
# default)
intervalLabel(1, 10, style = 1) # "(1,10]"
intervalLabel(1, 10, style = 2) # "<= 10"
intervalLabel(1, 10, style = 3) # "> 1"
intervalLabel(1, 10, style = 4) # "<= " "> 1" (vector of two elements!)
intervalLabel(1, 10, style = 5) # "<= 10" "> " (vector of two elements!)

# The same with left closed intervals:
right <- FALSE
intervalLabel(1, 10, right, style = 1) # "[1,10)"
intervalLabel(1, 10, right, style = 2) # "< 10"
intervalLabel(1, 10, right, style = 3) # ">= 1"
intervalLabel(1, 10, right, style = 4) # "< " ">= 1" (vector of two elements!)
intervalLabel(1, 10, right, style = 5) # "< 10" ">= " (vector of two elements!)

}
