% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invalid_zip.R
\name{invalid_zip}
\alias{invalid_zip}
\title{Path to (Invalid) Example Zip File}
\usage{
invalid_zip(...)
}
\arguments{
\item{\dots}{(segments of) relative path to the wanted zip file, starting
from "extdata/testdata/invalid-zips" in the package installation folder}
}
\description{
Path to (Invalid) Example Zip File
}
\examples{
# list available files
(available <- invalid_zip())
# Get full path to the first file
invalid_zip(available[1])
}
