% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kable.R
\name{kable_no_rows}
\alias{kable_no_rows}
\title{Print Data Frame as Markdown Table (Without Row Names by Default)}
\usage{
kable_no_rows(..., row.names = FALSE)
}
\arguments{
\item{\dots}{passed to \code{\link[knitr]{kable}}}

\item{row.names}{passed to \code{\link[knitr]{kable}}, default: \code{FALSE}}
}
\description{
Print Data Frame as Markdown Table (Without Row Names by Default)
}
