% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kable.R
\name{kable_translated}
\alias{kable_translated}
\title{Rename Data Frame Columns and Print as Markdown}
\usage{
kable_translated(x, ...)
}
\arguments{
\item{x}{x}

\item{\dots}{passed to \code{translate_columns}}
}
\description{
Rename Data Frame Columns and Print as Markdown
}
