% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown.R
\name{md_header}
\alias{md_header}
\title{Print Markdown Section Header}
\usage{
md_header(
  level,
  caption_key = "key?",
  caption = NULL,
  print = TRUE,
  msg = TRUE
)
}
\arguments{
\item{level}{level}

\item{caption_key}{caption_key}

\item{caption}{caption}

\item{print}{print}

\item{msg}{msg}
}
\description{
Print Markdown Section Header
}
