% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_berlin.R
\name{overwriteIfNotNA}
\alias{overwriteIfNotNA}
\title{Use Non-NA Values from Source Column in Target Column}
\usage{
overwriteIfNotNA(data, target_column, source_column)
}
\arguments{
\item{data}{data frame}

\item{target_column}{name of target column}

\item{source_column}{name of source column}
}
\description{
Overwrite the values in the target column with the values in the source
column at indices where the values in the source column are not NA
}
