% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_data_frames.R
\name{print.data_frame_diff}
\alias{print.data_frame_diff}
\title{Print Result of Data Frame Comparison}
\usage{
\method{print}{data_frame_diff}(x, ...)
}
\arguments{
\item{x}{object of class "data_frame_diff"}

\item{\dots}{currently not used}
}
\description{
Print Result of Data Frame Comparison
}
