% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_berlin.R
\name{printNumberOfNA}
\alias{printNumberOfNA}
\title{Print Number of NA Values in Given Column}
\usage{
printNumberOfNA(data, column, name = NULL)
}
\arguments{
\item{data}{data frame}

\item{column}{column name}

\item{name}{name of data}
}
\description{
Print Number of NA Values in Given Column
}
