% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_berlin.R
\name{printTableForColumn}
\alias{printTableForColumn}
\title{Print Result of table() for Given Column}
\usage{
printTableForColumn(data, column, name = NULL)
}
\arguments{
\item{data}{data frame}

\item{column}{column name}

\item{name}{name of data}
}
\description{
Print Result of table() for Given Column
}
