% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_actual_regrouping.R
\name{read_actual_regrouping}
\alias{read_actual_regrouping}
\title{Read and Filter "regroup_actual.csv"}
\usage{
read_actual_regrouping(
  name_actual,
  group = NULL,
  columns = NULL,
  as_list = TRUE
)
}
\arguments{
\item{name_actual}{Base name of file in config folder, default:
"regroup_actual". The file specifies: which regroupings are arcually to be
applied? What are the names of input and output columns?}

\item{group}{Name of column in \code{<name_actual>.csv} containing non-empty
fields for rows that are to be considered. \code{NULL} (the default) means
that all rows of the file are considered.}

\item{columns}{names of (input) columns that are to be regrouped. Only those
regroupings are performed that work on these columns or columns that are
created during the regrouping. By default \code{columns = NULL} all
regroupings are used (unless \code{group} is given).}

\item{as_list}{it \code{TRUE} (the default) the actual regrouping
configuration is returned as a list (as required by \code{doRegroupings}),
otherwise as a data frame.}
}
\description{
Read and Filter "regroup_actual.csv"
}
