% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doRegroupings.R
\name{regroup}
\alias{regroup}
\title{Assign Values to Groups of Values}
\usage{
regroup(
  x,
  assignments,
  ignore.case = NULL,
  to.factor = FALSE,
  to.numeric = TRUE
)
}
\arguments{
\item{x}{vector of values}

\item{assignments}{list of assignments of the form \<key\> = \<values\> with
\<values\> being a vector of elements to be looked up in \code{x} and to be
replaced with \<key\> in the output vector}

\item{ignore.case}{if \code{TRUE} the case is ignored when comparing values}

\item{to.factor}{if \code{TRUE} the new values are converted to
\code{factor}. The default is \code{FALSE}.}

\item{to.numeric}{if \code{TRUE} (the default!) and independent of
\code{to.factor} (!) the returned values are converted to numeric values if
all assigned (even though string) values "look like" numeric values, such
as "1", "2", "3.4", "5.67".}
}
\value{
vector with as many elements as there are elements in \code{x}. The
vector contains \<key\> at positions where the elements in \code{x} appeared
in the vector \<values\> of a \<key\> = \<values\> assignment of
\code{assignments}
}
\description{
Assign Values to Groups of Values
}
\examples{
regroup(c("A", "B", "C", "D"), assignments = list(
  "AB" = c("A", "B"),
  "CD" = c("C", "D")
))


x <- c("A", "B", "C", "D", "E", "A")
assignments <- list(
 "1" = c("A", "B"),
 "2" = c("C", "D")
)

regroup(x, assignments)

# to.factor is ignored...
regroup(x, assignments, to.factor = TRUE)

# ... unless to.numeric is FALSE!
regroup(x, assignments, to.factor = TRUE, to.numeric = FALSE)

}
