% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doRegroupings.R
\name{regroupedValues}
\alias{regroupedValues}
\title{Regroup Values According to Configuration}
\usage{
regroupedValues(
  values,
  config = NULL,
  labels = "labels1",
  to.factor = FALSE,
  to.numeric = TRUE,
  dbg = TRUE
)
}
\arguments{
\item{values}{vector of values}

\item{config}{configuration (list) describing how to regroup. If the list
contains an element \code{breaks} the function \code{\link{groupByBreaks}}
is called to group values together that belong to the same intervals that
are defined by the breaks. Otherwise the list must contain an element
\code{values} and an element of the name given in \code{labels} (default:
"labels1"). These are given to the function \code{\link{regroup}} that
performs a "value to label"-regrouping.}

\item{labels}{default: "labels1"}

\item{to.factor}{if \code{TRUE} the new values are converted to
\code{factor}. The default is \code{FALSE}.}

\item{to.numeric}{(default: \code{TRUE}, overrides \code{to.factor}!), passed
to \code{\link{regroup}}}

\item{dbg}{if \code{TRUE} (default) debug messages are shown}
}
\description{
Regroup Values According to Configuration
}
