% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regrouping_is_used.R
\name{regrouping_is_used}
\alias{regrouping_is_used}
\title{Which actual regroupings would be used?}
\usage{
regrouping_is_used(columns, actuals)
}
\arguments{
\item{columns}{vector of column names for which to check if they are subject
to regrouping}

\item{actuals}{list of elements \code{from} and \code{to}, as returned by
\code{kwb.prep:::read_actual_regrouping}}
}
\value{
vector of logical as long as \code{actuals}. Attribute \code{column}:
which columns would the data frame have after the regrouping?
}
\description{
Which of the actual regroupings would be used if \code{columns} were
available in a data frame
}
