% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_berlin.R
\name{removeRowsThatAreNaInColumn}
\alias{removeRowsThatAreNaInColumn}
\title{Remove Rows That are NA in Given Column}
\usage{
removeRowsThatAreNaInColumn(data, column, dbg = TRUE)
}
\arguments{
\item{data}{data frame}

\item{column}{column name}

\item{dbg}{it \code{TRUE} debug messages are shown}
}
\value{
\code{data} with rows removed that are \code{NA} in
\code{data[[column]]}
}
\description{
Remove Rows That are NA in Given Column
}
\examples{
df <- data.frame(a = c(1, NA, 3), b = c(11, 22, NA))
df

removeRowsThatAreNaInColumn(df, "a")
removeRowsThatAreNaInColumn(df, "b")

}
