% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_berlin.R
\name{replaceByCondition}
\alias{replaceByCondition}
\title{Replace Values in Column in Rows Matching Condition}
\usage{
replaceByCondition(df, file = NULL, group = NULL, config = NULL, dbg = TRUE)
}
\arguments{
\item{df}{data frame in which to do substitutions}

\item{file}{path to CSV file with columns \code{group},
\code{target_column}, \code{condition}, \code{replacement}}

\item{group}{group name. If given, only rows in \code{file} that have
this group name in column \code{group} are considered.}

\item{config}{optional. Data frame containing the configuration as being read
from \code{file}.}

\item{dbg}{if \code{TRUE} debug messages are shown}
}
\description{
Replace Values in Column in Rows Matching Condition
}
\examples{
# Create a very simple data frame
df <- data.frame(a = 1:3)

# Create a very simple configuration
config <- read.table(sep = ",", header = TRUE, text = c(
  "group,target,condition,replacement",
  "g1,a,a==2,22",
  "g2,a,a==3,33"
))

# Write the configuration to a temporary file
file <- tempfile()
write.csv(config, file)

# Apply all replacements configured in the file ...
replaceByCondition(df, file)

# ... or in the configuration
replaceByCondition(df, config = config)

# Apply selected replacements
replaceByCondition(df, file, group = "g1")
replaceByCondition(df, file, group = "g2")

}
