% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_berlin.R
\name{replaceUnlessNA}
\alias{replaceUnlessNA}
\title{Use Elements of Substitute at Indices Where Substitutes Are Not NA}
\usage{
replaceUnlessNA(x, substitute)
}
\arguments{
\item{x}{vector in which to substitute}

\item{substitute}{vector containing substitutions}
}
\description{
Use Elements of Substitute at Indices Where Substitutes Are Not NA
}
