% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_berlin.R
\name{reportNA}
\alias{reportNA}
\title{Count NA in a Column and Give a Message about It}
\usage{
reportNA(data, column, subject = "in data")
}
\arguments{
\item{data}{data frame}

\item{column}{name of column in \code{data}}

\item{subject}{value for placeholder \emph{subject} in output: "NAs
\emph{subject}: \emph{count_NA}"}
}
\description{
Count NA in a Column and Give a Message about It
}
