% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_column.R
\name{set_column}
\alias{set_column}
\title{Set Column}
\usage{
set_column(
  df,
  column,
  value = NULL,
  indices = NULL,
  from = NULL,
  must_exist = TRUE
)
}
\arguments{
\item{df}{data frame}

\item{column}{column}

\item{value}{value}

\item{indices}{row indices}

\item{from}{name of source column, optional}

\item{must_exist}{is column assumed to exist?}
}
\description{
Set Column
}
