% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_user_config_dir.R
\name{set_user_config_dir}
\alias{set_user_config_dir}
\title{Set the Path to the Config Folder}
\usage{
set_user_config_dir(path)
}
\arguments{
\item{path}{path to config folder containing configuration files}
}
\description{
Many functions in this package require configuration files. The idea behind
this is to separate (configurational) data from code. The code can the be
the same and only the configuration files need to be exchanged. With this
function you tell kwb.prep where to look for configuration files.
}
