% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_strings.R
\name{set_user_strings}
\alias{set_user_strings}
\title{Get List of User-Defined Text Constants}
\usage{
set_user_strings(x)
}
\arguments{
\item{x}{list of key = (character) value assignments}
}
\description{
Get List of User-Defined Text Constants
}
