% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop_text.R
\name{stop_text}
\alias{stop_text}
\title{Stop with Error Message Looked Up by Keyword}
\usage{
stop_text(...)
}
\arguments{
\item{\dots}{arguments passed to \code{\link{get_text}}}
}
\description{
Stop with Error Message Looked Up by Keyword
}
