% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_berlin.R
\name{stopOnDuplicates}
\alias{stopOnDuplicates}
\title{Stop If There Are Duplicates over given Columns}
\usage{
stopOnDuplicates(data, columns = names(data))
}
\arguments{
\item{data}{data frame}

\item{columns}{names of columns over which to look for duplicates. By
default, all columns in \code{data} are used.}
}
\description{
Stop If There Are Duplicates over given Columns
}
