% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unzip_archive.R
\name{unzip_archive}
\alias{unzip_archive}
\title{Unzip Archive}
\usage{
unzip_archive(zip_file, target_dir = tempdir(), flatten = TRUE, dbg = TRUE)
}
\arguments{
\item{zip_file}{path to archive file}

\item{target_dir}{path to target directory}

\item{flatten}{if \code{TRUE} (the default) all files in the archive are
unzipped directly into the target directory, independent from possible
folder structures within the archive}

\item{dbg}{whether or not to show debug messages}
}
\description{
Unzip Archive
}
