% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_filter_info.R
\name{write_filter_info}
\alias{write_filter_info}
\title{Write Information on Filtering to CSV files}
\usage{
write_filter_info(x, target_dir, prefix = deparse(substitute(x)), dbg = TRUE)
}
\arguments{
\item{x}{data frame as returned by \code{\link{apply_filters}}, with
attribute \code{filter_info} set.}

\item{target_dir}{path to directory into which to write csv files}

\item{prefix}{string by which to prefix all files}

\item{dbg}{whether or not to show debug messages}
}
\value{
\code{x}, unchanged, invisibly
}
\description{
Write Information on Filtering to CSV files
}
