% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown.R
\name{write_markdown_chapter}
\alias{write_markdown_chapter}
\title{Write a Markdown Chapter}
\usage{
write_markdown_chapter(x, caption_key = "key?", level = 3L, caption = NULL)
}
\arguments{
\item{x}{x}

\item{caption_key}{caption_key}

\item{level}{level}

\item{caption}{caption}
}
\description{
Write a Markdown Chapter
}
