% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_berlin.R
\name{writeStandardCsv}
\alias{writeStandardCsv}
\title{Write CSV File in a Standardised Manner}
\usage{
writeStandardCsv(x, file, ...)
}
\arguments{
\item{x}{data frame}

\item{file}{path to CSV file to be written}

\item{\dots}{additional arguments passed to \code{\link[utils]{write.table}}}
}
\description{
Write CSV File in a Standardised Manner
}
