% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{assign_objects}
\alias{assign_objects}
\title{Provide all Objects of kwb.prep in the Global Environment}
\usage{
assign_objects()
}
\description{
Provide all Objects of kwb.prep in the Global Environment
}
