% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assign_objects}
\alias{assign_objects}
\title{Assign all Objects of This Package in the Global Environment}
\usage{
assign_objects()
}
\description{
Assign all Objects of This Package in the Global Environment
}
