% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_text.R
\name{create_text_getter}
\alias{create_text_getter}
\title{Create a get_text() Function}
\usage{
create_text_getter(raw_strings)
}
\arguments{
\item{raw_strings}{list of string definitions (key = value) pairs}
}
\value{
a function that can be used to lookup the string constant(s)
}
\description{
Create a get_text() Function
}
\examples{
get_text <- create_text_getter(
  list(hello_en = "good morning", hello_de = "sch<oe>ne Gr<ue><ss>e")
)

get_text("hello_en")
get_text("hello_de")
#get_text("no_such_key") # error with clear error message

}
